validation on supplies.create/edit
no negative numbers on prices
2 decimal places accepted on prices
no negative numbers on quantity
error messages sent on respective views

+++ database/migrations/2022_08_04_023740_create_supplies_table.php
@@ -20,7 +20,8 @@ return new class extends Migration
             $table->string('name')->unique();
             $table->string('type');
             $table->string('brand')->nullable();
-            $table->integer('price');
+            // $table->integer('price');
+            $table->decimal('price', 5, 2);
             $table->integer('quantity');
             // $table->double('price', 8, 2);
         });
